Fix swarm tool registration to align with standardised approach
