+++ autogen/tools/dependency_injection.py
@@ -6,8 +6,7 @@ import functools
 import inspect
 import sys
 from abc import ABC
-from functools import wraps
-from typing import TYPE_CHECKING, Any, Callable, Iterable, Optional, TypeVar, Union, get_type_hints
+from typing import TYPE_CHECKING, Any, Callable, Optional, TypeVar, Union, get_type_hints
 
 from fast_depends import Depends as FastDepends
 from fast_depends import inject
@@ -15,6 +14,7 @@ from fast_depends.dependencies import model
 
 from ..agentchat import Agent
 from ..doc_utils import export_module
+from ..tools.function_utils import fix_staticmethod, remove_params
 
 if TYPE_CHECKING:
     from ..agentchat.conversable_agent import ConversableAgent
@@ -132,19 +132,14 @@ def _is_depends_param(param: inspect.Parameter) -> bool:
     )
 
 
-def _remove_params(func: Callable[..., Any], sig: inspect.Signature, params: Iterable[str]) -> None:
-    new_signature = sig.replace(parameters=[p for p in sig.parameters.values() if p.name not in params])
-    func.__signature__ = new_signature  # type: ignore[attr-defined]
-
-
 def _remove_injected_params_from_signature(func: Callable[..., Any]) -> Callable[..., Any]:
     # This is a workaround for Python 3.9+ where staticmethod.__func__ is accessible
     if sys.version_info >= (3, 9) and isinstance(func, staticmethod) and hasattr(func, "__func__"):
-        func = _fix_staticmethod(func)
+        func = fix_staticmethod(func)
 
     sig = inspect.signature(func)
     params_to_remove = [p.name for p in sig.parameters.values() if _is_context_param(p) or _is_depends_param(p)]
-    _remove_params(func, sig, params_to_remove)
+    remove_params(func, sig, params_to_remove)
     return func
 
 
@@ -188,20 +183,6 @@ def _string_metadata_to_description_field(func: Callable[..., Any]) -> Callable[
     return func
 
 
-def _fix_staticmethod(f: Callable[..., Any]) -> Callable[..., Any]:
-    # This is a workaround for Python 3.9+ where staticmethod.__func__ is accessible
-    if sys.version_info >= (3, 9) and isinstance(f, staticmethod) and hasattr(f, "__func__"):
-
-        @wraps(f.__func__)
-        def wrapper(*args: Any, **kwargs: Any) -> Any:
-            return f.__func__(*args, **kwargs)  # type: ignore[attr-defined]
-
-        wrapper.__name__ = f.__func__.__name__
-
-        f = wrapper
-    return f
-
-
 def _set_return_annotation_to_any(f: Callable[..., Any]) -> Callable[..., Any]:
     if inspect.iscoroutinefunction(f):
 
@@ -241,7 +222,7 @@ def inject_params(f: Callable[..., Any]) -> Callable[..., Any]:
     """
     # This is a workaround for Python 3.9+ where staticmethod.__func__ is accessible
     if sys.version_info >= (3, 9) and isinstance(f, staticmethod) and hasattr(f, "__func__"):
-        f = _fix_staticmethod(f)
+        f = fix_staticmethod(f)
 
     f = _string_metadata_to_description_field(f)
     f = _set_return_annotation_to_any(f)
