+++ autogen/tools/tool.py
@@ -3,10 +3,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import inspect
+import sys
 from typing import TYPE_CHECKING, Any, Callable, Optional, Union
 
 from ..doc_utils import export_module
-from ..tools.function_utils import get_function_schema
+from ..tools.function_utils import fix_staticmethod, get_function_schema, remove_params
 from .dependency_injection import ChatContext, get_context_params, inject_params
 
 if TYPE_CHECKING:
@@ -148,6 +149,20 @@ class Tool:
         schema = {"type": schema["type"], **schema["function"]}
 
         return schema
+
+    def hide_parameters(self, parameter_names: list[str]) -> None:
+        """Hide parameters from the tool's schema. Applicable for LLM registration.
+
+        Args:
+            parameter_names (list[str]): The names of the parameters to hide.
+        """
+        if sys.version_info >= (3, 9) and isinstance(self._func, staticmethod) and hasattr(self._func, "__func__"):
+            self._func = fix_staticmethod(self._func)
+
+        sig = inspect.signature(self._func)
+        params_to_remove = parameter_names
+
+        remove_params(self._func, sig, params_to_remove)
 
 
 @export_module("autogen.tools")
